move controllers tests directory, create board test
