+++ DonutsCoffees.Api.Tests/DonutsCoffees.Api.Tests.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="GamesTests\" />
+    <Folder Include="ControllersTests\" />
+  </ItemGroup>
 </Project>
