+++ DonutsCoffees.Api.Tests/GamesTests/BoardTest.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using NUnit.Framework;
+
+namespace DonutsCoffees.Api.Tests.GamesTests
+{
+    [TestFixture]
+    public class BoardTest
+    {
+        [Test]
+        public void ItReturnsEmptyBoard()
+        {
+          Board board = new Board();
+          AssertThat(board.empty == null);
+        }
+    }
+}
