+++ DonutsCoffees.Api/DonutsCoffees.Api.csproj
@@ -29,6 +29,13 @@
   <ItemGroup>
     <ProjectReference Include="..\DonutsCoffees.Api.Tests\DonutsCoffees.Api.Tests.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Games\" />
+    <Folder Include="Games\Definitions\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Remove="Games\Definitions\IPlayer.cs" />
+  </ItemGroup>
   <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
     <!-- Ensure Node.js is installed -->
     <Exec Command="node --version" ContinueOnError="true">
