+++ DonutsCoffees.Api/Games/Board.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace DonutsCoffees.Api.Games
+{
+    public class Board
+    {
+        public Board()
+        {
+        }
+    }
+}
