+++ DonutsCoffees.Api/Games/Definitions/IBoard.cs
@@ -1,0 +1,7 @@
+﻿using System;
+namespace DonutsCoffees.Api.Games
+{
+    public interface IBoard
+    {
+    }
+}
